ver 0.3

+++ SENSOR_FUSION/sensor_fusion_lib/include/measurement_package.h
@@ -9,7 +9,7 @@
 
 class MeasurementPackage {
 public:
-    long timestamp_;
+    float timestamp_;
 
     enum SensorType {
         LASER, RADAR
